Improve language selection by using Select2 library
For now, this only allows you to manually input each language. In the
future, once we have an API of sorts to give us all valid languages for
the chosen OCR engine, we can use that to restrict Select2 to only
allowing those languages, and also having autocomplete.

This PR also commits package-lock.json and removes yarn.lock which we
don't recommend using anymore. Also the dev-server doesn't appear to
work correctly, but it's not that important anyway, so we'll just
instruct devs to use 'npm run watch'.

Also removed the 'form-label' class from elements, which doesn't
do anything.

Bug: T280214