Ignore invalid languages for API requests
Instead of failing hard with an exception, we add a warning to the
response.
This is mainly useful for requests from wikisource, where we pass the
content language by default, but it might be unsupported.

Also make the engines return structured output and add code for
supporting warnings even in the UI, although it's currently impossible
to get any.

Bug: T285544