Return URL regex to requiring http prefix
This partially undoes the change in 78e773cb1c308a2b9ed01370ba01ad79f4363f97
and makes the http/https prefix required for image URLs.
This is because protocol-relative URLs are now being changed
to https in OcrController, so by the time the URL gets to the
EngineBase it's no longer relative.

Bug: T324740